Add files via upload
The primary objective of this fieldwork was to study and document the geological, 
Paleontological, sedimentary, Geomorphic and stratigraphic features of the 
Baripada beds focusing on their Mio Pliocene age. The aim was to identify and 
analyze fossil assemblages, to interpret the paleoenvironment and depositional 
settings of the formation. This work contributes to a deeper understanding of 
sedimentary processes, the stratigraphy of the region, and the palaeoecological 
conditions of the Mio Pliocene period.